+++ src/components/ItemList/index.ts
@@ -1,2 +1,2 @@
 export { default } from "./ItemList";
-export type { ItemAction } from "./ItemList";
+export type { ItemAction, ItemListHandle } from "./ItemList";